// vitals/core/src/models/product-card-inputs.model.ts
export interface ProductCardInputs {
  config?: string;
  productSku: string;
  productUrlTemplate: string;
  name: string;
  imageSrc: string;
  imageAlt: string;
  showPrice: string;
  showBadge: string;
  layout: string;
  cardLayout: string;
  theme: string;
  variant: string;
  variantKey: string;
}